+++ packages/core-js/actual/iterator/concat.js
@@ -1,5 +1,7 @@
 'use strict';
 var parent = require('../../stable/iterator/concat');
+require('../../modules/esnext.iterator.chunks');
 require('../../modules/esnext.iterator.concat');
+require('../../modules/esnext.iterator.windows');
 
 module.exports = parent;