+++ packages/core-js/actual/iterator/from.js
@@ -2,6 +2,7 @@
 var parent = require('../../stable/iterator/from');
 require('../../modules/es.promise');
 require('../../modules/esnext.iterator.constructor');
+require('../../modules/esnext.iterator.chunks');
 require('../../modules/esnext.iterator.dispose');
 require('../../modules/esnext.iterator.drop');
 require('../../modules/esnext.iterator.every');
@@ -16,5 +17,6 @@ require('../../modules/esnext.iterator.some');
 require('../../modules/esnext.iterator.take');
 require('../../modules/esnext.iterator.to-array');
 require('../../modules/esnext.iterator.to-async');
+require('../../modules/esnext.iterator.windows');
 
 module.exports = parent;