+++ packages/core-js/actual/iterator/index.js
@@ -3,6 +3,7 @@ var parent = require('../../stable/iterator');
 require('../../modules/es.object.create');
 require('../../modules/es.promise');
 require('../../modules/esnext.iterator.constructor');
+require('../../modules/esnext.iterator.chunks');
 require('../../modules/esnext.iterator.concat');
 require('../../modules/esnext.iterator.dispose');
 require('../../modules/esnext.iterator.drop');
@@ -18,6 +19,7 @@ require('../../modules/esnext.iterator.some');
 require('../../modules/esnext.iterator.take');
 require('../../modules/esnext.iterator.to-array');
 require('../../modules/esnext.iterator.to-async');
+require('../../modules/esnext.iterator.windows');
 require('../../modules/esnext.iterator.zip');
 require('../../modules/esnext.iterator.zip-keyed');