+++ packages/core-js/actual/iterator/zip-keyed.js
@@ -16,6 +16,8 @@ require('../../modules/es.iterator.some');
 require('../../modules/es.iterator.take');
 require('../../modules/es.iterator.to-array');
 require('../../modules/es.reflect.own-keys');
+require('../../modules/esnext.iterator.chunks');
+require('../../modules/esnext.iterator.windows');
 require('../../modules/esnext.iterator.zip-keyed');
 require('../../modules/web.dom-collections.iterator');