+++ tests/entries/unit.mjs
@@ -718,7 +718,9 @@ for (PATH of ['core-js-pure', 'core-js']) {
     ok(typeof load(NS, 'async-iterator/take') == 'function');
     ok(typeof load(NS, 'async-iterator/to-array') == 'function');
     ok(load(NS, 'function/metadata') === null);
+    ok(typeof load(NS, 'iterator/chunks') == 'function');
     ok(typeof load(NS, 'iterator/to-async') == 'function');
+    ok(typeof load(NS, 'iterator/windows') == 'function');
     ok(typeof load(NS, 'iterator/zip') == 'function');
     ok(typeof load(NS, 'iterator/zip-keyed') == 'function');
     ok(load(NS, 'symbol/metadata'));
@@ -779,9 +781,7 @@ for (PATH of ['core-js-pure', 'core-js']) {
     ok(typeof load(NS, 'iterator/as-indexed-pairs') == 'function');
     ok(typeof load(NS, 'iterator/indexed') == 'function');
     ok(load(NS, 'iterator/range')(1, 2).next().value === 1);
-    ok(typeof load(NS, 'iterator/chunks') == 'function');
     ok(typeof load(NS, 'iterator/sliding') == 'function');
-    ok(typeof load(NS, 'iterator/windows') == 'function');
     ok(load(NS, 'map/delete-all')(new Map(), 1, 2) === false);
     ok(load(NS, 'map/emplace')(new Map([[1, 2]]), 1, { update: it => it ** 2 }) === 4);
     ok(load(NS, 'map/every')(new Map([[1, 2], [2, 3], [3, 4]]), it => it % 2) === false);